* added particle pool (less memory consumption, less garbage collection, increase performance for a high number of particles) * optimized modifiers (a prepare function which is called once) * added editor and renderer for black holes

+++ src/main/java/de/hda/particles/renderer/BlackHoleRenderer.java
@@ -34,7 +34,7 @@ public class BlackHoleRenderer extends AbstractSelectable<BlackHole> implements 
 					Float gravityPointZ = new Float((Double) configuration.get(GravityPoint.POINT_Z));
 					Float gravity = new Float((Double) configuration.get(BlackHole.GRAVITY));
 					Float mass = new Float((Double) configuration.get(BlackHole.MASS));
-					Float eventHorizon = gravity * mass / BlackHole.EVENT_HORIZON_FACTOR;
+					Float eventHorizon = gravity * mass / BlackHole.EVENT_HORIZON_FACTOR.floatValue();
 					glPushMatrix();
 					glColor4f(1.0f, 0.6f, 0.0f, 0.3f);
 			        glTranslatef(gravityPointX, gravityPointY, gravityPointZ);
